Cleanup in progress TODO: Fix UserProperties
