+++ old/frontend-spinner.js
