+++ old/index.html
@@ -1,0 +1,70 @@
+<html>
+  <head>
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Darker+Grotesque:wght@900&display=swap" rel="stylesheet">
+    <link href="css/normal.css" type="text/css" rel="stylesheet">
+    <link href="css/login.css" type="text/css" rel="stylesheet">
+    <title>Akvario</title>
+    <link rel="icon" href="../public/resources/favicon.png">
+  </head>
+  <body>
+    <div class="login-form-wrapper">
+      <div class="bg-graphics"></div>
+      <div class="login-form">
+        <div class="login-form-content">
+          <h1>Welcome to the Akvario!</h1>
+          <div id="user-preview"></div>
+          <form id="FIXME">
+            <input type="text" id="username" name="username" placeholder="Enter username" minlength="2" maxlength="12" required>
+            <button type="submit">Join</button>
+          </form>
+        </div>
+        <div class="color-picker"><div class="color-picker-items"></div></div>
+      </div>
+    </div>
+
+      <div class="menuPopUp" id="popup">
+        <img type="image" id="menuBody" alt="menuBody" src="../public/resources/menuBody.svg">
+        <input type="image" class="speakers" alt="speakerIcon" id="speakers" src="../public/resources/speakerIcon.svg" onclick="muteUser()">
+      </div>
+
+        <div id="space">
+            <div class = "spinner">
+                <svg id = "spinner"
+                    xmlns="http://www.w3.org/2000/svg"
+                    width="93.436px"
+                    height="65.132px">
+                    <path
+                        d="M 34.803252,65.13249 C 15.581947,65.13249 0,50.552086 0,32.566248 0,14.580406 15.581947,4.8542449e-7 34.803252,0 c 16.915728,4.8542449e-7 50.208497,25.257473 57.95199,31.316697 0.907946,0.710463 0.906961,1.878831 -0.0021,2.58823 C 85.005598,39.950273 51.717508,65.13249 34.803371,65.13249 Z"
+                    />
+                </svg>
+            </div>
+        </div>
+
+        <script src="oldmain.js" type="module"></script>
+        <script src="user.js" type="module"></script>
+        <template id="userTemplate">
+            <div class="user-container">
+                <div class="body">
+                    <svg
+                            class="arrow"
+                            xmlns="http://www.w3.org/2000/svg"
+                            viewBox="0 0 11.328031 13.449599"
+                            height="20px"
+                            width="20px">
+                        <g transform="translate(-91.411618,-71.178162)">
+                            <g><path d="m 345.7462,274.54091 c 0.82399,3.92069 1.4126,7.88738 1.7626,11.87874 0.4234,4.82846 0.42341,12.70724 10e-6,17.5357 -0.31671,3.6118 -0.82884,7.20385 -1.53415,10.76031 -0.92943,4.68658 1.30847,6.46831 5.44756,4.07875 l 8.78623,-5.07241 a 226183.95,226183.95 149.99924 0 0 15.78977,-9.11651 l 9.67765,-5.58809 a 5.2634256,5.2634256 89.998346 0 0 -2.6e-4,-9.11639 l -9.91419,-5.7239 a 1153838.2,1153838.2 29.999328 0 0 -15.31587,-8.84238 l -8.94831,-5.16608 c -4.47177,-2.58166 -6.81263,-0.67902 -5.75104,4.37226 z" transform="scale(0.26458333)"/>
+                            </g>
+                        </g>
+                    </svg>
+                </div>
+                <h3 class="name" id="0_name"></h3>
+            </div>
+        </template>
+        <script src="https://unpkg.com/peerjs@1.2.0/dist/peerjs.min.js"></script>
+        <script src="oldmain.js" type="module"></script>
+        <script src="user.js" type="module"></script>
+        <script src="userClient.js"></script>
+        <script src="clientjs/socket.io.js"></script>
+  </body>
+</html>
