+++ old/routes/index.js
@@ -1,0 +1,6 @@
+import express from 'express'
+const router = express.Router();
+
+export default router.get('/', (req, res) => {
+    res.render('index'); 
+});
