+++ package.json
@@ -6,9 +6,10 @@
   "type": "module",
   "scripts": {
     "start": "node server.js",
-    "dev": "nodemon server.js -L"
+    "dev": "nodemon -e * app.js -L"
   },
   "dependencies": {
+    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "moment": "^2.29.1",
     "socket.io": "^3.1.2"
