--- routes/index.js
@@ -1,6 +1,0 @@
-import express from 'express'
-const router = express.Router();
-
-export default router.get('/', (req, res) => {
-    res.render('index'); 
-});
