+++ scripts/sockets.js
@@ -1,0 +1,8 @@
+import {UserProperties} from "./user.js";
+
+export function handleConnection(socket, userData){
+    socket.emit('connected')
+
+    socket.on('hello', () => console.log('world'));
+}
+
