+++ scripts/user.js
@@ -1,122 +1,91 @@
-/*
-Here all users are stored and deleted
-freeIDs is ids not being used but has been used.
-*/
-let users = [];
-let freeID = [];
 
 
+export class UserProperties{
+    constructor() {
+        this.users = {};
+        this.userProperties = {
+            position: {},
+            rotation: {},
+            name:     {},
+            color:    {}
+        }
+    }
 
-/*return a new id for new users*/
-function newID(){
+    // Returns the user with the given id.
+    get = id => {
+        if (!this.users.hasOwnProperty(id)) return undefined;
 
-  /*if there is ids free then give one to the new user and delete it from freeID*/
-  if(freeID.length != 0){
-    let id = freeID[0];
-    freeID.shift();
-    return id;
-  }
+        const properties = {id: id};
+        Object.keys(this.userProperties).forEach(prop =>
+            properties[prop] = this.userProperties[prop][id]);
 
-  /*if there is no free IDs give the new user the length of users as id */
-  return users.length;
+        return properties;
+    }
+
+    // Adds a user to collection, Returns the id.
+    add = (id, name, color, position, rotation) => {
+        // Check if it exists already...
+        if (this.exists(id))
+            console.warn(`user with id ${id} already existed! Did you mean to override?`);
+
+        // ..Otherwise, add it.
+        else {
+            const properties = this.userProperties;
+            const user = this.users[id] = {
+                get position() {return properties.position(id)},
+                get rotation() {return properties.rotation(id)},
+                get name()     {return properties.name(id)},
+                get color()    {return properties.color(id)},
+
+                set position(position) {properties.position[id] = position},
+                set rotation(rotation) {properties.rotation[id] = rotation},
+                set name(name)   {properties.name[id]  = name},
+                set color(color) {properties.color[id] = color}
+            };
+
+            user.name  = name;
+            user.color = color;
+            user.position = position;
+            user.rotation = rotation;
+        }
+        return this.get(id);
+    }
+
+    get positions() {return this.userProperties.position}
+    get rotations() {return this.userProperties.rotation}
+    get names()     {return this.userProperties.name}
+    get colors()    {return this.userProperties.color}
+
+    // Does the user with the given id exist?
+    exists = id => this.users.hasOwnProperty(id);
+
+    // Removes the user with the given id from collection
+    remove = id =>{
+        if(this.exists(id)){
+            Object.keys(this.userProperties).forEach(prop => delete this.userProperties[prop][id]);
+            delete this.users[id];
+        }
+
+        else console.warn(`user with id ${id} did not exist and could not be deleted!`);
+    }
+
+    toJSON = () => {
+        const allUsers = [];
+        for (const user in this.users){
+            allUsers.push(this.get(user));
+        }
+        return JSON.stringify(allUsers);
+    }
 }
 
+const allUsers = new UserProperties();
+const freeIDs = [];
+let previousID = 0;
+let nextID = () => ++previousID;
+const initialPosition = {top:10, left:10};
+const initialRotation = 0;
 
-
-/*deletes the user who left*/
-function deleteID(id){
-  const index = findIndexID(users, id);
-  console.log(`---${index}---`);
-
-  //if there is a index with the id delete it and put it in freeID
-  if(index > -1){
-    users.splice(index, 1);
-    freeID.push(id);
-  }
-}
-
-
-
-//user template
-function initUser(id){
-  this.name = "";
-  this.id = id;
-  this.color = [0, 0, 0];
-  this.pos = [0, 0];
-  this.rad = 0;
-  this.generateColors = ()=>{
-    for(let i = 0; i < this.color.length; i++)
-      this.color[i] = Math.floor(Math.random() * 256);
-  }
-}
-
-
-
-//create new user
-function createUser(id){ 
-
-  //adds a new user from the user template
-  users.push(new initUser(id));
-  const i = users.length - 1;
-
-  //this should be changed
-  users[i].name = "name";
-
-  //generates colors
-  users[i].generateColors();
-
-  //initial position in percentage
-  users[i].pos[0] = Math.floor(Math.random() * (90 - 10) + 10);
-  users[i].pos[1] = Math.floor(Math.random() * (90 - 10) + 10);
-
-  //generateBody(id);
-
-  //when createUser is done it returns the id that can be used in other places
-  return id;
-}
-
-
-
-//look in /public/js/userClient.js
-function findIndexID(arr, id){
-  for(let i = 0; i < arr.length; i++){
-    if(arr[i].id == id)
-      return i;
-  }
-  return -1;
-}
-
-
-
-//shows the new user in the server console
-function showNewProp(i){
-  if(i != -1)
-    console.log(users[i]);
-  else
-  console.log("nej");
-  
-}
-
-
-
-//shows all the active and free ids 
-function showAll(){
-  console.log("--Active IDs--")
-  for(let i = 0; i < users.length; i++)
-    console.log(users[i].id)
-  console.log("--Free IDs--");
-  for(let i = 0; i < freeID.length; i++)
-    console.log(freeID[i]);
-}
-
-
-// exports the functions to be used outside this js file
-module.exports = {
-  createUser,
-  newID,
-  deleteID,
-  showNewProp,
-  showAll,
-  users,
-  findIndexID
+export function createUser(name, color){
+    const id = freeIDs.pop() || nextID();
+    return allUsers.add(id, name, color, initialPosition, initialRotation);
 }
