+++ server.js
@@ -2,9 +2,9 @@ import express from 'express';
 import path from 'path';
 import {createServer} from 'http';
 import * as socket_io from 'socket.io';
-import {UserCollection, colorPicker} from './public/js/user.js'
-import indexRouter from './routes/index.js';
-import spin from './scripts/backend-spinner.js';
+import {UserCollection, colorPicker} from './old/user.js'
+import indexRouter from './old/routes/index.js';
+import spin from './old/backend-spinner.js';
 
 //express and path modules
 const app = express();
